+++ src/types/logger.d.ts
@@ -1,5 +1,7 @@
 import { App } from 'vue'
 
+export type LogLevel = 'debug' | 'info' | 'warn' | 'error' | 'log'
+
 /**
  * @class
  * @example
@@ -27,6 +29,7 @@ export declare class VueLogger {
   /**
    * Applies log options to this Logger.
    * The provided options will be merged with existing options already applied.
+   *
    * @param options {LoggerOptions}
    */
   apply (options: LoggerOptions): void
@@ -34,37 +37,42 @@ export declare class VueLogger {
    * Perform debug level logging.
    *
    * Only performed if LoggerOptions.level is one of: debug
+   *
    * @param args
    */
-  debug (...args: any): void
+  debug (...args: any): Promise<void>
   /**
    * Perform info level logging.
    *
    * Only performed if LoggerOptions.level is one of: debug, info
+   *
    * @param args
    */
-  info (...args: any): void
+  info (...args: any): Promise<void>
   /**
    * Perform warning level logging.
    *
    * Only performed if LoggerOptions.level is one of: debug, info, warn
+   *
    * @param args
    */
-  warn (...args: any): void
+  warn (...args: any): Promise<void>
   /**
    * Perform error level logging.
    *
    * Only performed if LoggerOptions.level is one of: debug, info, warn, error
+   *
    * @param args
    */
-  error (...args: any): void
+  error (...args: any): Promise<void>
   /**
    * Perform generic (console.log) level logging.
    *
    * Performed if Logger.Options.level is one of: debug, info, warn, error, log
+   *
    * @param args
    */
-  log (...args: any): void
+  log (...args: any): Promise<void>
   /**
    * Returns whether logging is enabled.
    */
@@ -72,12 +80,13 @@ export declare class VueLogger {
   /**
    * Returns the currently applied logging level (one of: debug, info, warn, error, log).
    */
-  level (): string
+  level (): LogLevel
   install (app: App): void
 }
 
 /**
  * Create a VueLogger instance that can be used on a Vue app.
+ *
  * @param options - {@link LoggerOptions}
  * @example
  * const logger = createLogger({
@@ -103,6 +112,7 @@ export function useLogger (): VueLogger
 
 /**
  * Options for logging functionality.
+ *
  * @interface
  */
 export interface LoggerOptions {
@@ -117,7 +127,17 @@ export interface LoggerOptions {
   /**
    * @field level {string} the logging level (one of: debug, info, warn, error)
    */
-  level?: string
+  level?: LogLevel
+  /**
+   * @field callerInfo {boolean} whether information about the caller function should be included
+   * @see {@link CallerInfo}
+   */
+  callerInfo?: boolean
+  /**
+   * @field provide a custom formatted string for the log message prefix (preceeds the log arguments)
+   * @param event {@link LogEvent}
+   */
+  prefixFormat? (event: Readonly<Pick<LogEvent, 'level' | 'caller'>>): string
   /**
    * @field beforeHooks {LoggerHook[]} hooks invoked before a statement is logged, can be used to alter log arguments (use carefully)
    */
@@ -130,25 +150,41 @@ export interface LoggerOptions {
 
 /**
  * Provides custom logic to be executed during a logging event.
+ *
  * @interface
  * @example
  * const customHook: LoggerHook = {
  *  run (event: LogEvent) {
  *    // event.level is the log level invoked (e.g. debug, info, warn, error)
  *    // event.argumentArray is the array of arguments which were passed to the logging method
+ *    // event.caller contains the caller function information (e.g. function name, file name, line number), if LoggerOptions.callerInfo is true
  *  }
  * }
  */
 export interface LoggerHook {
-  run (event: LogEvent): void
+  run (event: LogEvent): void | Promise<void>
   install? (options: LoggerOptions): void
   props?: { [key: string]: any }
 }
 
 /**
- * Contains information for a log invocation (log level, array of arguments).
+ * Contains information for a log invocation (log level, array of arguments, caller function info).
+ *
+ * @interface
  */
 export interface LogEvent {
-  level: string
+  level: LogLevel
   argumentArray: any[]
+  caller?: CallerInfo
+}
+
+/**
+ * Information about the caller function which invoked the logger.
+ *
+ * @interface
+ */
+export interface CallerInfo {
+  fileName?: string
+  functionName?: string
+  lineNumber?: string
 }