+++ tests/logging.test.ts
@@ -12,44 +12,44 @@ describe('logging: levels', () => {
     jest.resetAllMocks()
   })
 
-  it('debug logs to console.debug', () => {
+  it('debug logs to console.debug', async () => {
     console.debug = jest.fn()
-    VueE.config.globalProperties.$log.debug('test')
-    VueE.config.globalProperties.$log.debug('test', testObject)
-    expect(console.debug).toHaveBeenCalledWith('debug | ', 'test')
-    expect(console.debug).toHaveBeenCalledWith('debug | ', 'test', testObject)
+    await VueE.config.globalProperties.$log.debug('test')
+    await VueE.config.globalProperties.$log.debug('test', testObject)
+    expect(console.debug).toHaveBeenCalledWith('[DEBUG]', 'test')
+    expect(console.debug).toHaveBeenCalledWith('[DEBUG]', 'test', testObject)
   })
 
-  it('info logs to console.info', () => {
+  it('info logs to console.info', async () => {
     console.info = jest.fn()
-    VueE.config.globalProperties.$log.info('test')
-    VueE.config.globalProperties.$log.info('test', testObject)
-    expect(console.info).toHaveBeenCalledWith('info | ', 'test')
-    expect(console.info).toHaveBeenCalledWith('info | ', 'test', testObject)
+    await VueE.config.globalProperties.$log.info('test')
+    await VueE.config.globalProperties.$log.info('test', testObject)
+    expect(console.info).toHaveBeenCalledWith('[INFO]', 'test')
+    expect(console.info).toHaveBeenCalledWith('[INFO]', 'test', testObject)
   })
 
-  it('warn logs to console.warn', () => {
+  it('warn logs to console.warn', async () => {
     console.warn = jest.fn()
-    VueE.config.globalProperties.$log.warn('test')
-    VueE.config.globalProperties.$log.warn('test', testObject)
-    expect(console.warn).toHaveBeenCalledWith('warn | ', 'test')
-    expect(console.warn).toHaveBeenCalledWith('warn | ', 'test', testObject)
+    await VueE.config.globalProperties.$log.warn('test')
+    await VueE.config.globalProperties.$log.warn('test', testObject)
+    expect(console.warn).toHaveBeenCalledWith('[WARN]', 'test')
+    expect(console.warn).toHaveBeenCalledWith('[WARN]', 'test', testObject)
   })
 
-  it('error logs to console.error', () => {
+  it('error logs to console.error', async () => {
     console.error = jest.fn()
-    VueE.config.globalProperties.$log.error('test')
-    VueE.config.globalProperties.$log.error('test', testObject)
-    expect(console.error).toHaveBeenCalledWith('error | ', 'test')
-    expect(console.error).toHaveBeenCalledWith('error | ', 'test', testObject)
+    await VueE.config.globalProperties.$log.error('test')
+    await VueE.config.globalProperties.$log.error('test', testObject)
+    expect(console.error).toHaveBeenCalledWith('[ERROR]', 'test')
+    expect(console.error).toHaveBeenCalledWith('[ERROR]', 'test', testObject)
   })
 
-  it('log logs to console.log', () => {
+  it('log logs to console.log', async () => {
     console.log = jest.fn()
-    VueE.config.globalProperties.$log.log('test')
-    VueE.config.globalProperties.$log.log('test', testObject)
-    expect(console.log).toHaveBeenCalledWith('log | ', 'test')
-    expect(console.log).toHaveBeenCalledWith('log | ', 'test', testObject)
+    await VueE.config.globalProperties.$log.log('test')
+    await VueE.config.globalProperties.$log.log('test', testObject)
+    expect(console.log).toHaveBeenCalledWith('[LOG]', 'test')
+    expect(console.log).toHaveBeenCalledWith('[LOG]', 'test', testObject)
   })
 
 })
@@ -60,12 +60,12 @@ describe('logging: unsupported', () => {
     jest.resetAllMocks()
   })
 
-  it('unsupported console function logs to console.log', () => {
+  it('unsupported console function logs to console.log', async () => {
     console.warn = undefined
     console.log = jest.fn()
     const logger = createLogger()
-    logger.warn('test')
-    expect(console.log).toHaveBeenCalledWith('warn | ', 'test')
+    await logger.warn('test')
+    expect(console.log).toHaveBeenCalledWith('[WARN]', 'test')
   })
 
 })
@@ -76,12 +76,12 @@ describe('logging: hooks', () => {
     jest.resetAllMocks()
   })
 
-  it('invokes before hooks', () => {
+  it('invokes before hooks', async () => {
     const hook = {
       run: jest.fn()
     }
     const logger = createLogger({ beforeHooks: [hook] })
-    logger.log('test')
+    await logger.log('test')
     expect(hook.run).toHaveBeenCalledWith(
       expect.objectContaining({
         level: 'log',
@@ -90,12 +90,12 @@ describe('logging: hooks', () => {
     )
   })
 
-  it('invokes after hooks', () => {
+  it('invokes after hooks', async () => {
     const hook = {
       run: jest.fn()
     }
     const logger = createLogger({ afterHooks: [hook] })
-    logger.log('test')
+    await logger.log('test')
     expect(hook.run).toHaveBeenCalledWith(
       expect.objectContaining({
         level: 'log',
@@ -104,14 +104,14 @@ describe('logging: hooks', () => {
     )
   })
 
-  it('handles hook run failure', () => {
+  it('handles hook run failure', async () => {
     console.warn = jest.fn()
     const hook = {
       run: jest.fn()
     }
     hook.run.mockImplementation(() => { throw Error('Test') })
     const logger = createLogger({ beforeHooks: [hook] })
-    logger.log('test')
+    await logger.log('test')
     expect(console.warn).toHaveBeenCalledWith(
       'LoggerHook run failure',
       expect.anything()
@@ -126,20 +126,20 @@ describe('logging: disabled', () => {
     jest.resetAllMocks()
   })
 
-  it('does not write to console', () => {
+  it('does not write to console', async () => {
     console.log = jest.fn()
     const logger = createLogger({ enabled: false })
-    logger.log('test')
+    await logger.log('test')
     expect(console.log).not.toHaveBeenCalled()
   })
 
-  it('does not invoke hooks', () => {
+  it('does not invoke hooks', async () => {
     console.log = jest.fn()
     const hook = {
       run: jest.fn()
     }
     const logger = createLogger({ enabled: false, beforeHooks: [hook] })
-    logger.log('test')
+    await logger.log('test')
     expect(hook.run).not.toHaveBeenCalled()
   })
 
@@ -151,21 +151,67 @@ describe('logging: console disabled', () => {
     jest.resetAllMocks()
   })
 
-  it('does not write to console', () => {
+  it('does not write to console', async () => {
     console.log = jest.fn()
     const logger = createLogger({ consoleEnabled: false })
-    logger.log('test')
+    await logger.log('test')
     expect(console.log).not.toHaveBeenCalled()
   })
 
-  it('does invoke hooks', () => {
+  it('does invoke hooks', async () => {
     console.log = jest.fn()
     const hook = {
       run: jest.fn()
     }
     const logger = createLogger({ consoleEnabled: false, beforeHooks: [hook] })
-    logger.log('test')
+    await logger.log('test')
     expect(hook.run).toHaveBeenCalled()
+  })
+
+})
+
+describe('logging: prefix format', () => {
+
+  afterEach(() => {
+    jest.resetAllMocks()
+  })
+
+  it('uses custom format when provided', async () => {
+    console.debug = jest.fn()
+    const logger = createLogger({ prefixFormat: ({ level }) => `[${level.toUpperCase()}]` })
+    await logger.debug('test')
+    expect(console.debug).toHaveBeenCalledWith('[DEBUG]', 'test')
+  })
+
+})
+
+describe('logging: caller info', () => {
+
+  afterEach(() => {
+    jest.resetAllMocks()
+  })
+
+  afterAll(() => {
+    jest.restoreAllMocks()
+  })
+
+  async function doLog(logger) {
+    await logger.debug('test')
+  }
+
+  it('logs caller info', async () => {
+    console.debug = jest.fn()
+    const logger = createLogger({ callerInfo: true })
+    await doLog(logger)
+    expect(console.debug).toHaveBeenCalledWith('[DEBUG] [logging.test.ts:doLog:199]', 'test')
+  })
+
+  it('does not log caller info when stack unavailable', async () => {
+    const mockErrorStack = jest.spyOn(window.Error, 'prepareStackTrace') as any
+    mockErrorStack.mockImplementation(() => undefined)
+    const logger = createLogger({ callerInfo: true })
+    await doLog(logger)
+    expect(console.debug).toHaveBeenCalledWith('[DEBUG]', 'test')
   })
 
 })